Update induction_cooking.htm
The eddy currents produced by an induction cooker are not the result of accelerating the entire metal atom itself, but of free charges like electrons able to transmit electric currents. Additionally, metals like steel are not actually composed of metal molecules. A molecule is a discrete collection of atoms held together by chemical bonds, like H2O. So I think it's safer just to leave it more vaguely phrased.