Add `<meta name='color-scheme'>`
The [`color-scheme` meta tag][0] provides a signal to the browser about
what color schemes the website supports it. Without it, browsers default
to light mode, which means that viewing the page before any CSS has
loaded will produce a white background even if the device is in dark
mode.

The `color-scheme` meta tag activates the browser's awareness of the
OS's color preferences, thus it will choose a dark background for
unstyled content, preventing the flash of white.

[0]: https://html.spec.whatwg.org/multipage/semantics.html#meta-color-scheme