philosophy.htm: Add width, height, refactor ethics
In addition to adding dimensions to the images, I also refactored the
icons that link to the XXIIVV Ethics page: Moving the link to a separate
`<p>` means that the icons' alt text won't be overshadowed by the link
label.

See 4df2006f866b7bf986c0211e5afb6f419ba20bcd for an example of that.